# openapi.json
{
  "components": {
    "schemas": {
...
      "AppealCreate": {
        "properties": {
          "decision_id": {
            "format": "uuid",
            "title": "Decision Id",
            "type": "string"
          },
          "decision_snapshot": {
            "additionalProperties": true,
            "title": "Decision Snapshot",
            "type": "object"
          },
          "decision_type": {
            "enum": [
              "qa",
              "deliverable",
              "payout",
              "credential",
              "reputation"
            ],
            "title": "Decision Type",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "decision_type",
          "decision_id",
          "decision_snapshot"
        ],
        "title": "AppealCreate",
        "type": "object"
      },
      "AppealResolve": {
        "properties": {
          "decision": {
            "enum": [
              "UPHELD",
              "OVERTURNED",
              "PARTIALLY_OVERTURNED"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "maxLength": 4000,
            "minLength": 20,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reason"
        ],
        "title": "AppealResolve",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearningContentSection": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body"
        ],
        "title": "LearningContentSection",
        "type": "object"
      },
      "LearningModuleCompleteRequest": {
        "properties": {
          "evidence_summary": {
            "maxLength": 2000,
            "minLength": 20,
            "title": "Evidence Summary",
            "type": "string"
          }
        },
        "required": [
          "evidence_summary"
        ],
        "title": "LearningModuleCompleteRequest",
        "type": "object"
      },
      "LearningModuleView": {
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "completion_evidence": {
            "title": "Completion Evidence",
            "type": "string"
          },
          "content_sections": {
            "items": {
              "$ref": "#/components/schemas/LearningContentSection"
            },
            "title": "Content Sections",
            "type": "array"
          },
          "estimated_minutes": {
            "title": "Estimated Minutes",
            "type": "integer"
          },
          "exercise_brief": {
            "title": "Exercise Brief",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ordinal",
          "title",
          "summary",
          "estimated_minutes",
          "content_sections",
          "exercise_brief",
          "completion_evidence",
          "completed"
        ],
        "title": "LearningModuleView",
        "type": "object"
      },
      "LearningPathView": {
        "properties": {
          "enrolled": {
            "title": "Enrolled",
            "type": "boolean"
          },
          "estimated_hours": {
            "title": "Estimated Hours",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/LearningModuleView"
            },
            "title": "Modules",
            "type": "array"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisites",
            "type": "array"
          },
          "progress_percent": {
            "title": "Progress Percent",
            "type": "integer"
          },
          "skill_outcomes": {
            "items": {
              "type": "string"
            },
            "title": "Skill Outcomes",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "level",
          "estimated_hours",
          "skill_outcomes",
          "prerequisites",
          "modules",
          "enrolled",
          "progress_percent",
          "status"
        ],
        "title": "LearningPathView",
        "type": "object"
      },
...
      "ReputationEventCreate": {
        "properties": {
          "dimension": {
            "enum": [
              "reliability",
              "delivered_quality",
              "collaboration",
              "verified_skill"
            ],
            "title": "Dimension",
            "type": "string"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "evidence_type": {
            "enum": [
              "client_acceptance",
              "lead_review",
              "appeal_resolution",
              "coordinator_investigation"
            ],
            "title": "Evidence Type",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "student_user_id": {
            "format": "uuid",
            "title": "Student User Id",
            "type": "string"
          },
          "value": {
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "student_user_id",
          "project_id",
          "dimension",
          "value",
          "evidence_type",
          "evidence_id"
        ],
        "title": "ReputationEventCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PraxisAI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/appeals": {
      "get": {
        "operationId": "list_appeals_api_v1_appeals_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Appeals Api V1 Appeals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Appeals",
        "tags": [
          "appeals and reputation"
        ]
      },
      "post": {
        "operationId": "create_appeal_api_v1_appeals_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Appeal Api V1 Appeals Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Appeal",
        "tags": [
          "appeals and reputation"
        ]
      }
    },
...
    "/api/v1/learning/modules/{module_id}/complete": {
      "post": {
        "operationId": "complete_api_v1_learning_modules__module_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningModuleCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningPathView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/paths": {
      "get": {
        "operationId": "paths_api_v1_learning_paths_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearningPathView"
                  },
                  "title": "Response Paths Api V1 Learning Paths Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Paths",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/paths/{path_id}/enroll": {
      "post": {
        "operationId": "enroll_api_v1_learning_paths__path_id__enroll_post",
        "parameters": [
          {
            "in": "path",
            "name": "path_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Path Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningPathView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enroll",
        "tags": [
          "learning"
        ]
      }
    },
...
    "/api/v1/ops/appeals/{appeal_id}/resolve": {
      "post": {
        "operationId": "resolve_appeal_api_v1_ops_appeals__appeal_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "appeal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appeal Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealResolve"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Appeal Api V1 Ops Appeals  Appeal Id  Resolve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Appeal",
        "tags": [
          "appeals and reputation"
        ]
      }
    },
...
    "/api/v1/ops/reputation-events": {
      "post": {
        "operationId": "create_reputation_event_api_v1_ops_reputation_events_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReputationEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Reputation Event Api V1 Ops Reputation Events Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Reputation Event",
        "tags": [
          "appeals and reputation"
        ]
      }
    },
...
    "/api/v1/students/me/reputation-events": {
      "get": {
        "operationId": "my_reputation_events_api_v1_students_me_reputation_events_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "praxis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Praxis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response My Reputation Events Api V1 Students Me Reputation Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My Reputation Events",
        "tags": [
          "appeals and reputation"
        ]
      }
    },
...
  }
}